% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlsMclust.R
\name{mvls.Mclust}
\alias{mvls.Mclust}
\title{mvls.Mclust}
\usage{
mvls.Mclust(data, imp.method = "mean", G = 1:9)
}
\arguments{
\item{data}{A dataset (more than two longitudinal mesurements)}

\item{imp.method}{Type od pre-imputation. Mclust doesn't use missing value. Possible "mean" or "locf".}

\item{G}{The vector of possible cluster}
}
\value{
$data It's the data-set with imputation

$Mclust It's a Mclust object

$sd.1 It contains the sd for each data imputed at single imputation method
}
\description{
Takes a longitudinal dataset and impute missing value with Mclust machine learning-based method.
}
\author{
Lorenzo Querci <lorenzo.querci@studio.unibo.it>
}
