% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlspam.R
\name{mvls.pam}
\alias{mvls.pam}
\title{mvls.pam}
\usage{
mvls.pam(data, imp.method = "mean", krange = 2:12, scaling = T)
}
\arguments{
\item{data}{A dataset (more than two longitudinal mesurements).}

\item{imp.method}{Type od pre-imputation. pam and pamk don't use missing value. Possible "mean" or "locf".}

\item{krange}{Range for k in pamk (fpc package) function.}

\item{scaling}{TRUE/FALSE value (default T). Scale data before clustering.}
}
\value{
$data It's the data-set with imputation.

$pam It's a pam object

$sd.1 It contains the sd for each data imputed at single imputation method
}
\description{
Takes a longitudinal dataset and impute missing value with pam (k by pamk) machine learning-based method.
}
\author{
Lorenzo Querci <lorenzo.querci@studio.unibo.it>
}
