% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlsprint.R
\name{mvls.print}
\alias{mvls.print}
\title{mvlsprint}
\usage{
mvls.print(data, d, method = "h", varmatrix = F, kmax = 10)
}
\arguments{
\item{data}{A dataset (more than two longitudinal mesurements)}

\item{d}{percentage of change between two-sided mesurements to consider it bigger, smaller or the same, It useful to built the var.matrix}

\item{method}{it represent the type of machine learning algorithm. 'k' for k-mean and 'h' for hierical}

\item{varmatrix}{TRUE/FALSE value (default it's F). It permit to see cluster for var.matrix}

\item{kmax}{It's the maximum number of cluster for k-mean. Default setting it's 10. It's the elbow k-mean method.}
}
\value{
Plot
}
\description{
Takes a longitudinal dataset and give you graphic plot to choose number of cluster
}
\author{
Lorenzo Querci <lorenzo.querci@studio.unibo.it>
}
