% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlspvclust.R
\name{mvls.pvclust}
\alias{mvls.pvclust}
\title{mvls.pvclust}
\usage{
mvls.pvclust(mvls, data = "data", nboot = 1000)
}
\arguments{
\item{mvls}{A mvls object after h cluster analysis}

\item{data}{Type of data analysis. "data" for analysis on real data after imputtion or "var.matrix" for analysis on vari.matrix.}

\item{nboot}{Number of boot operation.}
}
\value{
$pvclust It's a pvclust object.
}
\description{
Hierical cluster control with pvclust application
}
\author{
Lorenzo Querci <lorenzo.querci@studio.unibo.it>
}
