% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvls.R
\name{mvls}
\alias{mvls}
\title{mvls}
\usage{
mvls(data, d = 0.1, method = "k", cluster = 6, nstart = 20,
  pre.imp = F, imp.method = "mean")
}
\arguments{
\item{data}{A dataset (more than two longitudinal mesurements)}

\item{d}{Percentage of change between two-sided mesurements to consider it bigger, smaller or the same, It useful to built the var.matrix.}

\item{method}{It represent the type of machine learning algorithm. 'k' for k-mean and 'h' for hierical.}

\item{cluster}{It's the number of cluster. Default setting it's 6. It depends on number of longitudinal mesurements. It could be use mvls.print to decide best cluster number.}

\item{nstart}{It is the nstart setting of function k-mean. Defualt it'20. Not requested for 'h' method.}

\item{pre.imp}{TRUE/FALSE (default F). It permit to pre-impute data to built the vari.matrix, It could be reduce cluster with only missing value.}

\item{imp.method}{It's the type of pre-imputation. Defaul it's 'mean', but there is also 'locf' possibility.}
}
\value{
$data It's the data-set with imputation.

$cluster It's the cluster matrix.

$matrix It's the vari.matrix.

$sd.1 It contains the sd for each data imputed at single imputation method. Different from sd.2.

$vari.matrix It's the variation matrix

$data.norm It's the imputation dataset normalized
}
\description{
Takes a longitudinal dataset and impute missing value with a machine learning-based method.
}
\author{
Lorenzo Querci <lorenzo.querci@studio.unibo.it>
}
