% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlsboot.R
\name{mvlsboot}
\alias{mvlsboot}
\title{mvlsboot}
\usage{
mvlsboot(data, d = 0.1, method = "k", cluster = 6, nstart = 20,
  pre.imp = F, imp.method = "mean", boot = 10)
}
\arguments{
\item{data}{A dataset (more than two longitudinal mesurements)}

\item{d}{It is the percentage of change between two-sided mesurements to consider it bigger, smaller or the same, It useful to built the var.matrix}

\item{method}{It represent the type of machine learning algorithm. 'k' for k-mean and 'h' for hierical}

\item{cluster}{It's the number of cluster. Default setting it's 6. It depends on number of longitudinal mesurements. It could be use mvls.print to decide best cluster number.}

\item{nstart}{It is the nstart setting of function k-mean. Defualt it'20. Not requested for 'h' method.}

\item{pre.imp}{TRUE/FALSE (default F). It permit to pre-impute data to built the vari.matrix, It could be reduce cluster with only missing value.}

\item{imp.method}{It's the type of pre-imputation. Defaul it's 'mean', but there is also 'locf' possibility}

\item{boot}{Impute many times to reduce error from cluster imputation. 'low' for 5 imputation, 'medium' for 10 imputation and 'high' for 15 imputation.}
}
\value{
$data It's the data-set with imputation.

$sd.2 It contain the sd for each data imputed at multiple imputation method. Different from sd.1.

$db.boot It's a list of all imputed dataset (n=boot).
}
\description{
Takes a longitudinal dataset and impute missing value with a machine learning-based method many time.
}
\author{
Lorenzo Querci <lorenzo.querci@studio.unibo.it>
}
