% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualdiagmvls.R
\name{visualdiagmvls}
\alias{visualdiagmvls}
\title{visualdiagmvls}
\usage{
visualdiagmvls(mvls, method = "multiple", cluster, norm = F)
}
\arguments{
\item{mvls}{A object class mvls}

\item{method}{Indicate if all cluster plot together or single one (need to select it). "single" or "multiple"}

\item{cluster}{It's the number of clustet to plot. Number of pattern, max is the k decided for mvls function}

\item{norm}{Permit to plot from normalized data. TRUE/FALSE value}
}
\description{
Takes a mvls object and give a plot about cluster distribution
}
\author{
Lorenzo Querci <lorenzo.querci@studio.unibo.it>
}
